got SamplePlugin working as a system extension
Had difficulties finding a way to add a directory of jars to classpath
of an intellij run configuration. As a result, created a new task in the
application build.gradle.kts file that adds a runtime dependency
whenever the task is invoked. Added a new run configuration in intellij
that calls the new task, which then invokes the kotlin action that runs
the application.